Get installation : modification votes type Long --> int

+++ src/main/java/claurendeau/hackqc/algo/Backend/dto/InstallationDTO.java
@@ -2,5 +2,5 @@ package claurendeau.hackqc.algo.Backend.dto;
 
 public record InstallationDTO(Long id, String name,
                               String type, String description,
-                              Long votePour, Long voteContre) {
+                              int votePour, int voteContre) {
 }
